ci: test not organized import
