+++ packages/core/src/index.ts
@@ -1,7 +1,14 @@
+// organize-imports-ignore
 // Export a default Cache - we might recode this to be able to specify custom provider
 export { MemoryCache as Cache } from "ts-method-cache";
+import { MethodCacheService } from "ts-method-cache";
+const CacheService = new MethodCacheService();
+export { CacheService };
+
 export * from "./application";
+export * from "./unpackedapplication";
 export * from "./core";
+
 // Models
 export * from "./models/aclmodel";
 export * from "./models/coremodel";
@@ -10,11 +17,13 @@ export * from "./models/ident";
 export * from "./models/ownermodel";
 export * from "./models/rolemodel";
 export * from "./models/user";
+
 // Queues
 export * from "./queues/filequeue";
 export * from "./queues/memoryqueue";
 export * from "./queues/pubsubservice";
 export * from "./queues/queueservice";
+
 // Services
 export * from "./services/asyncevents";
 export * from "./services/authentication";
@@ -35,14 +44,14 @@ export * from "./services/oauth";
 export * from "./services/resource";
 export * from "./services/service";
 export * from "./services/version";
+
 // Stores
 export * from "./stores/aggregator";
 export * from "./stores/file";
 export * from "./stores/mapper";
 export * from "./stores/memory";
 export * from "./stores/store";
-export * from "./stores/webdaql/query";
-export * from "./unpackedapplication";
+
 // Utils
 export * from "./utils/abstractdeployer";
 export * from "./utils/context";
@@ -52,6 +61,5 @@ export * from "./utils/logger";
 export * from "./utils/serializers";
 export * from "./utils/session";
 export * from "./utils/waiter";
-export { CacheService };
-import { MethodCacheService } from "ts-method-cache";
-const CacheService = new MethodCacheService();
+
+export * from "./stores/webdaql/query";
